refactor: split SmartContractService into focused modules
- Create SmartContractDeployer for WASM deployment logic
- Create SmartContractInvoker for contract method invocation
- Create SmartContractReader for state reading and simulation
- Keep SmartContractService as facade delegating to new modules
- Maintains backward compatible public API

Closes #898
